+++ autogpt_platform/frontend/src/app/(platform)/library/agents/[id]/loading.tsx
@@ -1,3 +1,3 @@
-import { AgentRunsLoading } from "./components/NewAgentLibraryView/components/sidebar-lists/AgentRunsLoading";
+import { AgentRunsLoading } from "./components/NewAgentLibraryView/components/other/AgentRunsLoading";
 
 export default AgentRunsLoading;
